Rename as conversation-paused
